+++ blueprints/flexberry-model/ModelBlueprint.ts
@@ -23,6 +23,7 @@ class SortedPair{
 export default class ModelBlueprint {
   model: string;
   serializerAttrs: string;
+  offlineSerializerAttrs: string;
   parentModelName: string;
   parentClassName: string;
   className: string;
@@ -30,38 +31,46 @@ export default class ModelBlueprint {
   name: string;
   needsAllModels: string;
   needsAllEnums: string;
+  needsAllObjects: string;
   lodashVariables: {};
   constructor(blueprint, options) {
     let modelsDir = path.join(options.metadataDir, "models");
     if (!options.file) {
       options.file = options.entity.name + ".json";
     }
-    let modelFile = path.join(modelsDir, options.file);
-    let content: string = stripBom(fs.readFileSync(modelFile, "utf8"));
-    let model: metadata.Model = JSON.parse(content);
+    let model: metadata.Model = ModelBlueprint.loadModel(modelsDir, options.file);
     this.parentModelName = model.parentModelName;
     this.parentClassName = model.parentClassName;
     this.className = model.className;
     this.serializerAttrs = this.getSerializerAttrs(model);
+    this.offlineSerializerAttrs = this.getOfflineSerializerAttrs(model);
     this.projections = this.getJSForProjections(model, modelsDir);
     this.model = this.getJSForModel(model);
     this.name = options.entity.name;
     this.needsAllModels = this.getNeedsAllModels(modelsDir);
-    this.needsAllEnums = this.getNeedsAllEnums(path.join(options.metadataDir, "enums"));
+    this.needsAllEnums = this.getNeedsTransforms(path.join(options.metadataDir, "enums"));
+    this.needsAllObjects = this.getNeedsTransforms(path.join(options.metadataDir, "objects"));
     let modelLocales = new ModelLocales(model, modelsDir, "ru");
     this.lodashVariables = modelLocales.getLodashVariablesProperties();
   }
 
-  getNeedsAllEnums(enumsDir: string): string {
-    let listEnums = fs.readdirSync(enumsDir);
-    let enums: string[] = [];
-    for (let e of listEnums) {
+  static loadModel(modelsDir: string, modelFileName: string): metadata.Model {
+    let modelFile = path.join(modelsDir, modelFileName);
+    let content = stripBom(fs.readFileSync(modelFile, "utf8"));
+    let model: metadata.Model = JSON.parse(content);
+    return model;
+  }
+
+  getNeedsTransforms(dir: string): string {
+    let list = fs.readdirSync(dir);
+    let transforms: string[] = [];
+    for (let e of list) {
       let pp: path.ParsedPath = path.parse(e);
       if (pp.ext != ".json")
         continue;
-      enums.push(`    'transform:${pp.name}'`);
+      transforms.push(`    'transform:${pp.name}'`);
     }
-    return enums.join(",\n");
+    return transforms.join(",\n");
   }
 
   getNeedsAllModels(modelsDir: string): string {
@@ -87,7 +96,21 @@ export default class ModelBlueprint {
     if(attrs.length===0){
       return "";
     }
-    return "    "+attrs.join(",\n    ");
+    return "      "+attrs.join(",\n      ");
+  }
+
+  getOfflineSerializerAttrs(model: metadata.Model): string {
+    let attrs: string[] = [];
+    for (let belongsTo of model.belongsTo) {
+      attrs.push(belongsTo.name + ": { serialize: 'id', deserialize: 'records' }");
+    }
+    for (let hasMany of model.hasMany) {
+      attrs.push(hasMany.name + ": { serialize: 'ids', deserialize: 'records' }");
+    }
+    if (attrs.length === 0) {
+      return "";
+    }
+    return "      " + attrs.join(",\n      ");
   }
 
   getJSForModel(model: metadata.Model): string {
@@ -96,7 +119,15 @@ export default class ModelBlueprint {
     let templateHasMany = lodash.template("<%=name%>: DS.hasMany('<%=relatedTo%>', { inverse: <%if(inverse){%>'<%=inverse%>'<%}else{%>null<%}%>, async: false })");
     let attr: metadata.DSattr;
     for (attr of model.attrs) {
-      attrs.push(`${attr.name}: DS.attr('${attr.type}')`);
+      let comment = "";
+      if (!attr.stored) {
+        comment =
+          "/**\n" +
+          TAB + TAB + "Non-stored property.\n\n" +
+          TAB + TAB + `@property ${attr.name}\n` +
+          TAB + "*/\n" + TAB;
+      }
+      attrs.push(`${comment}${attr.name}: DS.attr('${attr.type}')`);
       if (attr.notNull) {
         if (attr.type === "date") {
           validations.push(attr.name + ": { datetime: true }");
@@ -104,6 +135,27 @@ export default class ModelBlueprint {
           validations.push(attr.name + ": { presence: true }");
         }
       }
+      if (attr.stored)
+        continue;
+      let methodToSetNotStoredProperty =
+        "/**\n" +
+        TAB + TAB + "Method to set non-stored property.\n" +
+        TAB + TAB + "Please, use code below in model class (outside of this mixin) otherwise it will be replaced during regeneration of models.\n" +
+        TAB + TAB + `Please, implement '${attr.name}Compute' method in model class (outside of this mixin) if you want to compute value of '${attr.name}' property.\n\n` +
+        TAB + TAB + `@method _${attr.name}Compute\n` +
+        TAB + TAB + "@private\n" +
+        TAB + TAB + "@example\n" +
+        TAB + TAB + TAB + "```javascript\n" +
+        TAB + TAB + TAB + `_${attr.name}Changed: Ember.on('init', Ember.observer('${attr.name}', function() {\n` +
+        TAB + TAB + TAB + TAB + `Ember.run.once(this, '_${attr.name}Compute');\n` +
+        TAB + TAB + TAB + "}))\n" +
+        TAB + TAB + TAB + "```\n" +
+        TAB + "*/\n" +
+        TAB + `_${attr.name}Compute: function() {\n` +
+        TAB + TAB + `let result = (this.${attr.name}Compute && typeof this.${attr.name}Compute === 'function') ? this.${attr.name}Compute() : null;\n` +
+        TAB + TAB + `this.set('${attr.name}', result);\n` +
+        TAB + "}";
+      attrs.push(methodToSetNotStoredProperty);
     }
     let belongsTo: metadata.DSbelongsTo;
     for (belongsTo of model.belongsTo) {
@@ -118,24 +170,25 @@ export default class ModelBlueprint {
     if(validations.length===0){
       validationsFunc="";
     }
-    validationsFunc = "getValidations: function () {\n" +
-    TAB + TAB + "let parentValidations = this._super();\n" +
-    TAB + TAB + "let thisValidations = {\n" +
-    validationsFunc + TAB + TAB + "};\n" +
-    TAB + TAB + "return Ember.$.extend(true, {}, parentValidations, thisValidations);\n" +
-    TAB + "}";
-    let initFunction = "init: function () {\n" +
-    TAB + TAB + "this.set('validations', this.getValidations());\n" +
-    TAB + TAB + "this._super.apply(this, arguments);\n" +
-    TAB + "}";
+    validationsFunc =
+      "getValidations: function () {\n" +
+      TAB + TAB + "let parentValidations = this._super();\n" +
+      TAB + TAB + "let thisValidations = {\n" +
+      validationsFunc + TAB + TAB + "};\n" +
+      TAB + TAB + "return Ember.$.extend(true, {}, parentValidations, thisValidations);\n" +
+      TAB + "}";
+    let initFunction =
+      "init: function () {\n" +
+      TAB + TAB + "this.set('validations', this.getValidations());\n" +
+      TAB + TAB + "this._super.apply(this, arguments);\n" +
+      TAB + "}";
     attrs.push(validationsFunc, initFunction);
     return TAB + attrs.join(",\n" + TAB);
   }
 
   joinProjHasMany(detailHasMany: metadata.ProjHasMany, modelsDir: string, level: number): SortedPair {
     let hasManyAttrs: SortedPair[] = [];
-    let modelFile = path.join(modelsDir, detailHasMany.relatedTo + ".json");
-    let hasManyModel: metadata.Model = JSON.parse(stripBom(fs.readFileSync(modelFile, "utf8")));
+    let hasManyModel: metadata.Model = ModelBlueprint.loadModel(modelsDir, detailHasMany.relatedTo + ".json");
     let hasManyProj = lodash.find(hasManyModel.projections, function(pr: metadata.ProjectionForModel) { return pr.name === detailHasMany.projectionName; });
     if (hasManyProj) {
       for (let attr of hasManyProj.attrs) {
@@ -222,8 +275,7 @@ export default class ModelBlueprint {
       }
       for (let hasMany of proj.hasMany) {
         let hasManyAttrs: SortedPair[] = [];
-        let modelFile = path.join(modelsDir, hasMany.relatedTo + ".json");
-        let detailModel: metadata.Model = JSON.parse(stripBom(fs.readFileSync(modelFile, "utf8")));
+        let detailModel: metadata.Model = ModelBlueprint.loadModel(modelsDir, hasMany.relatedTo + ".json");
         projName = hasMany.projectionName;
         let detailProj = lodash.find(detailModel.projections, function(pr: metadata.ProjectionForModel) { return pr.name === projName; });
         if (detailProj) {