+++ blueprints/flexberry-model/index.js
@@ -4,8 +4,8 @@
 "use strict";
 var ModelBlueprint_1 = require('./ModelBlueprint');
 var lodash = require('lodash');
-var Blueprint = require('ember-cli/lib/models/blueprint');
-var AddonBlueprint_1 = require('../flexberry-addon/AddonBlueprint');
+var path = require('path');
+var CommonUtils_1 = require('../flexberry-common/CommonUtils');
 module.exports = {
     description: 'Generates an ember-data model for flexberry.',
     availableOptions: [
@@ -15,9 +15,27 @@ module.exports = {
     supportsAddon: function () {
         return false;
     },
+    _files: null,
+    files: function () {
+        if (this._files) {
+            return this._files;
+        }
+        var modelsDir = path.join(this.options.metadataDir, "models");
+        if (!this.options.file) {
+            this.options.file = this.options.entity.name + ".json";
+        }
+        var model = ModelBlueprint_1.default.loadModel(modelsDir, this.options.file);
+        if (!model.offline) {
+            this._files = CommonUtils_1.default.getFilesForGeneration(this, function (v) { return v === "__root__/mixins/regenerated/serializers/__name__-offline.js"; });
+        }
+        else {
+            this._files = CommonUtils_1.default.getFilesForGeneration(this);
+        }
+        return this._files;
+    },
     afterInstall: function (options) {
         if (this.project.isEmberCLIAddon()) {
-            AddonBlueprint_1.default.install(options, ["model", "serializer"]);
+            CommonUtils_1.default.installFlexberryAddon(options, ["model", "serializer"]);
         }
     },
     /**
@@ -38,9 +56,11 @@ module.exports = {
             model: modelBlueprint.model,
             projections: modelBlueprint.projections,
             serializerAttrs: modelBlueprint.serializerAttrs,
+            offlineSerializerAttrs: modelBlueprint.offlineSerializerAttrs,
             name: modelBlueprint.name,
             needsAllModels: modelBlueprint.needsAllModels,
-            needsAllEnums: modelBlueprint.needsAllEnums // for use in files\tests\unit\serializers\__name__.js
+            needsAllEnums: modelBlueprint.needsAllEnums,
+            needsAllObjects: modelBlueprint.needsAllObjects // for use in files\tests\unit\serializers\__name__.js
         }, modelBlueprint.lodashVariables);
     }
 };