+++ tests/dummy/app/locales/ru/translations.js
@@ -432,6 +432,10 @@ Ember.$.extend(true, translations, {
               'caption': 'Пример скачивания файлов со списка',
               'title': ''
             },
+            'object-list-view-resize': {
+              'caption': 'Разметка столбцов',
+              'title': ''
+            }
           },
           'flexberry-simpledatetime': {
             'caption': 'flexberry-simpledatetime',
@@ -782,8 +786,15 @@ Ember.$.extend(true, translations, {
         'selected-rows': {
           'caption': 'Flexberry-objectlistview. Выбранные строки'
         },
+<<<<<<< HEAD
         'downloading-files-from-olv-list': {
           'caption': 'Flexberry-objectlistview. Скачивание файлов со списка'
+=======
+        'object-list-view-resize': {
+          'caption': 'Flexberry-objectlistview. Разметка столбцов',
+          'button-сaption': 'Добавить',
+          'title': ''
+>>>>>>> develop
         }
       },
       'flexberry-simpleolv': {